wavemc_backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-decouple==3.8
PyJWT==2.8.0
mysql-connector-python==8.0.33
email-validator==2.0.0
passlib==1.7.4
bcrypt==4.0.1
fastapi-mail==1.4.1
beautifulsoup4==4.12.2
aiosmtplib==2.0.2